#9 Bangladeshi tree draw

+++ solutions/01-raida/Age of Villagers/arc.cs
@@ -5,8 +5,8 @@ namespace Age_of_Villagers
     class arc : IShape
     {
         private readonly Point point;
-        private readonly float radius, startAngle, sweepAngle;
-        public arc(Point p,float radius, float startAngle,float sweepAngle)
+        private readonly int radius, startAngle, sweepAngle;
+        public arc(Point p,int radius, int startAngle,int sweepAngle)
         {
             point = p;
             this.radius = radius;
@@ -15,7 +15,7 @@ namespace Age_of_Villagers
         }
         public void draw(Graphics g, Pen pen)
         {
-            g.DrawArc(pen, point.X, point.Y, radius, radius, startAngle, sweepAngle);
+            g.DrawArc(pen, point.X,point.Y, radius,radius, startAngle, sweepAngle);
         }
     }
 }
